requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.49.2
idna==3.11
pillow==12.2.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
requests==2.33.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
